modify UserServiceImpl

+++ src/main/java/edu/university/program/service/impl/UserServiceImpl.java
@@ -1,10 +1,12 @@
 package edu.university.program.service.impl;
 
+import edu.university.program.exception.NullEntityReferenceException;
 import edu.university.program.model.User;
 import edu.university.program.repository.UserRepository;
 import edu.university.program.service.UserService;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -20,25 +22,41 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User create(User user) {
-        return userRepository.save(user);
+        try {
+            return userRepository.save(user);
+        } catch (IllegalArgumentException e) {
+            throw new NullEntityReferenceException("User cannot be 'null'");
+        }
     }
 
     @Override
     public User update(User user) {
-        User oldUser = readById(user.getId());
-        return userRepository.save(user);
+        if (user != null) {
+            User oldUser = readById(user.getId());
+            if (oldUser != null) {
+                return userRepository.save(user);
+            }
+        }
+        throw new NullEntityReferenceException("User cannot be 'null'");
     }
 
     @Override
     public User readById(long id) {
         Optional<User> optional = userRepository.findById(id);
-        return optional.get();
+        if (optional.isPresent()) {
+            return optional.get();
+        }
+        throw new EntityNotFoundException("User with id " + id + " not found");
     }
 
     @Override
     public void delete(long id) {
         User user = readById(id);
-        userRepository.delete(user);
+        if (user != null) {
+            userRepository.delete(user);
+        } else {
+            throw new EntityNotFoundException("User with id " + id + " not found");
+        }
     }
 
     @Override
